Pull _annotate_span_for_generation_message into AugmentedLLM
