+++ src/mcp_agent/workflows/evaluator_optimizer/evaluator_optimizer.py
@@ -1,8 +1,10 @@
 import contextlib
 from enum import Enum
+from opentelemetry import trace
 from typing import Callable, List, Optional, Type, TYPE_CHECKING
 from pydantic import BaseModel, Field
 
+from mcp_agent.tracing.semconv import GEN_AI_AGENT_NAME
 from mcp_agent.workflows.llm.augmented_llm import (
     AugmentedLLM,
     MessageParamT,
@@ -157,103 +159,113 @@ class EvaluatorOptimizerLLM(AugmentedLLM[MessageParamT, MessageT]):
         request_params: RequestParams | None = None,
     ) -> List[MessageT]:
         """Generate an optimized response through evaluation-guided refinement"""
-        refinement_count = 0
-        response = None
-        best_response = None
-        best_rating = QualityRating.POOR
-        self.refinement_history = []
+        tracer = self.context.tracer or trace.get_tracer("mcp-agent")
+        with tracer.start_as_current_span(
+            f"EvaluatorOptimizerLLM.{self.name}.generate"
+        ) as span:
+            span.set_attribute(GEN_AI_AGENT_NAME, self.agent.name)
+            self._annotate_span_for_generation_message(span, message)
 
-        # Initial generation
-        async with contextlib.AsyncExitStack() as stack:
-            if isinstance(self.optimizer, Agent):
-                await stack.enter_async_context(self.optimizer)
-            response = await self.optimizer_llm.generate(
-                message=message,
-                request_params=request_params,
-            )
+            if request_params:
+                self._annotate_span_with_request_params(span, request_params)
 
-        best_response = response
+            refinement_count = 0
+            response = None
+            best_response = None
+            best_rating = QualityRating.POOR
+            self.refinement_history = []
 
-        while refinement_count < self.max_refinements:
-            logger.debug("Optimizer result:", data=response)
-
-            # Evaluate current response
-            eval_prompt = self._build_eval_prompt(
-                original_request=str(message),
-                current_response="\n".join(str(r) for r in response)
-                if isinstance(response, list)
-                else str(response),
-                iteration=refinement_count,
-            )
-
-            evaluation_result = None
-            async with contextlib.AsyncExitStack() as stack:
-                if isinstance(self.evaluator, Agent):
-                    await stack.enter_async_context(self.evaluator)
-
-                evaluation_result = await self.evaluator_llm.generate_structured(
-                    message=eval_prompt,
-                    response_model=EvaluationResult,
-                    request_params=request_params,
-                )
-
-            # Track iteration
-            self.refinement_history.append(
-                {
-                    "attempt": refinement_count + 1,
-                    "response": response,
-                    "evaluation_result": evaluation_result,
-                }
-            )
-
-            logger.debug("Evaluator result:", data=evaluation_result)
-
-            # Track best response (using enum ordering)
-            if evaluation_result.rating.value > best_rating.value:
-                best_rating = evaluation_result.rating
-                best_response = response
-                logger.debug(
-                    "New best response:",
-                    data={"rating": best_rating, "response": best_response},
-                )
-
-            # Check if we've reached acceptable quality
-            if (
-                evaluation_result.rating.value >= self.min_rating.value
-                or not evaluation_result.needs_improvement
-            ):
-                logger.debug(
-                    f"Acceptable quality {evaluation_result.rating.value} reached",
-                    data={
-                        "rating": evaluation_result.rating.value,
-                        "needs_improvement": evaluation_result.needs_improvement,
-                        "min_rating": self.min_rating.value,
-                    },
-                )
-                break
-
-            # Generate refined response
-            refinement_prompt = self._build_refinement_prompt(
-                original_request=str(message),
-                current_response="\n".join(str(r) for r in response)
-                if isinstance(response, list)
-                else str(response),
-                feedback=evaluation_result,
-                iteration=refinement_count,
-            )
-
+            # Initial generation
             async with contextlib.AsyncExitStack() as stack:
                 if isinstance(self.optimizer, Agent):
                     await stack.enter_async_context(self.optimizer)
-
                 response = await self.optimizer_llm.generate(
-                    message=refinement_prompt,
+                    message=message,
                     request_params=request_params,
                 )
 
-            refinement_count += 1
+            best_response = response
 
-        return best_response
+            while refinement_count < self.max_refinements:
+                logger.debug("Optimizer result:", data=response)
+
+                # Evaluate current response
+                eval_prompt = self._build_eval_prompt(
+                    original_request=str(message),
+                    current_response="\n".join(str(r) for r in response)
+                    if isinstance(response, list)
+                    else str(response),
+                    iteration=refinement_count,
+                )
+
+                evaluation_result = None
+                async with contextlib.AsyncExitStack() as stack:
+                    if isinstance(self.evaluator, Agent):
+                        await stack.enter_async_context(self.evaluator)
+
+                    evaluation_result = await self.evaluator_llm.generate_structured(
+                        message=eval_prompt,
+                        response_model=EvaluationResult,
+                        request_params=request_params,
+                    )
+
+                # Track iteration
+                self.refinement_history.append(
+                    {
+                        "attempt": refinement_count + 1,
+                        "response": response,
+                        "evaluation_result": evaluation_result,
+                    }
+                )
+
+                logger.debug("Evaluator result:", data=evaluation_result)
+
+                # Track best response (using enum ordering)
+                if evaluation_result.rating.value > best_rating.value:
+                    best_rating = evaluation_result.rating
+                    best_response = response
+                    logger.debug(
+                        "New best response:",
+                        data={"rating": best_rating, "response": best_response},
+                    )
+
+                # Check if we've reached acceptable quality
+                if (
+                    evaluation_result.rating.value >= self.min_rating.value
+                    or not evaluation_result.needs_improvement
+                ):
+                    logger.debug(
+                        f"Acceptable quality {evaluation_result.rating.value} reached",
+                        data={
+                            "rating": evaluation_result.rating.value,
+                            "needs_improvement": evaluation_result.needs_improvement,
+                            "min_rating": self.min_rating.value,
+                        },
+                    )
+                    break
+
+                # Generate refined response
+                refinement_prompt = self._build_refinement_prompt(
+                    original_request=str(message),
+                    current_response="\n".join(str(r) for r in response)
+                    if isinstance(response, list)
+                    else str(response),
+                    feedback=evaluation_result,
+                    iteration=refinement_count,
+                )
+
+                async with contextlib.AsyncExitStack() as stack:
+                    if isinstance(self.optimizer, Agent):
+                        await stack.enter_async_context(self.optimizer)
+
+                    response = await self.optimizer_llm.generate(
+                        message=refinement_prompt,
+                        request_params=request_params,
+                    )
+
+                refinement_count += 1
+
+            return best_response
 
     async def generate_str(
         self,
